Separate "list php" logic for further reuse, and finish the work for unix
